# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Workshops API",
        "description": "REST API для управления мастер-классами, площадками и регистрациями. Вариант 15.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
...
        "/venues": {
            "get": {
                "tags": [
                    "Площадки"
                ],
                "summary": "Список площадок",
                "operationId": "b3d0a1c57452d0d30bba9b40560f2d18",
                "responses": {
                    "200": {
                        "description": "Список площадок с пагинацией"
                    }
                }
            },
            "post": {
                "tags": [
                    "Площадки"
                ],
                "summary": "Создать площадку (только админ)",
                "operationId": "0bbda601ee6beea68a0b232049f2d9e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "capacity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Новая студия"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "ул. Пушкина, д. 10"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "hourly_fee": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Площадка создана"
                    },
                    "403": {
                        "description": "Нет прав"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/venues/{id}": {
            "get": {
                "tags": [
                    "Площадки"
                ],
                "summary": "Просмотр площадки",
                "operationId": "b3b661a990ba6616b68e7bab9c723951",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные площадки"
                    },
                    "404": {
                        "description": "Не найдено"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Площадки"
                ],
                "summary": "Удалить площадку (только админ)",
                "operationId": "8be672420720a9436f5d19173758794a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Удалено"
                    },
                    "403": {
                        "description": "Нет прав"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Площадки"
                ],
                "summary": "Обновить площадку (только админ)",
                "operationId": "08b2d10cc70933d4154d1d97dd1eb228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Обновлённая студия"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновлено"
                    },
                    "403": {
                        "description": "Нет прав"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum",
                "description": "Вставьте токен в формате: Bearer {token}"
            }
        }
    },
...
}